sockets: Convert tcp client to struct+impl
Many approaches have been tried to achieve the solution reached.

The problem was that connect and send_message functions should be
independent. That means that the user should be able to connect
once and send multiple messages.

This approach had the problem that the tcp stream variable needed
to be shared among the connect and send_message methods, so it
needed to be part of the struct.
The fact that Rust does not have default values for struct members
made that the user had to initialize the tcp stream in order to
instantiate the struct. This was a problem because I wanted to
abstract the tcp stream from the user, so she only needed to know
host and port. To fix that issue many things were tried until this
final solution was reached.

The solutions were:

  1. Use TcpBuilder from net2 crate

This solution tried to give a default (unitialized) value to the
tcp stream given that we can change the tcp stream variable and
use TcpBuilder without starting a connection yet, just using the
associated function TcpBuilder::new_v4() and pass it when
instantiating the client struct.

This was still a problem because TcpBuilder sucks a lot and it does
not allow you to extract the TcpStream from the TcpBuilder multiple
times (convert from TcpBuilder to TcpStream). So the problem was that
the struct had the a variable of type TcpBuilder, but when the
TcpBuilder connect() method was called, the TcpStream variable inside
the TcpBuilder was extracted (and left) and when the send_message
method was called and self.tcp_builder.to_tcp_stream() was called, it
failed because the tcp stream inside the tcp builder was already
consumed by the connect.

This problem left us with the same issue we had before. We need
a tcp stream that needs to be shared among impl methods, but we only
get its value after instantiating the TCP client.

TcpBuilder from net2 crate allowed us to give a default (empty) value
to the tcp variable in the struct, but it could not be converted to
a tcp stream and passed among methods.

  2. Use builder pattern

This solution tried to hide the tcp stream variable from the user by
giving a ClientBuilder struct to the user that internally instantiates
the Client struct.

It did not work because when the builder was instantiated, we could
pass only host and port, but eventually, before the finalize() - when
the Client struct is instantiated - is called, the TcpStream must have
been passed. This left us in the same position as before, just delaying
a little bit the use of TcpStream and overengineering the code.

  3. Use an associated function

The final solution fixed the issue by joining the struct instantiation
and the connection to the server from the client. This solution could
have been implemented using the builder pattern, but the use of the
builder pattern was not necessary.

Using an associate function - new() - to the Client struct was an
intermediate point between having complete separation between struct
methods. There is some dependency, given that new() calls connect(),
but the user can expect that to happen in a socket library. Furthermore,
the associated function new() could be renamed to connect() to make
things clearer.

In this case, the associate function instantiates the Client struct
while calling connect, so the tcp stream is created and associated to
the struct variable that will be shared among all struct methods.